requirements.txt
torch==2.0
pandas==1.5.3
pdb2sql==0.5.3
openfold2==0.0.1
